Refactor `TestResult` component to make code more manageable.
